strlcpy y strnstr

+++ ft_memset.c
@@ -1,4 +1,6 @@
-void *ft_memset(void *s, int c, unsigned int n)
+#include "libft.h"
+
+void *ft_memset(void *s, int c, size_t n)
 {
 	unsigned int i;
